requirements.txt
fastapi
uvicorn
python-dotenv
tqdm
SQLAlchemy
pandas
httpx
google-auth-oauthlib
google-api-python-client
PyJWT
langchain-core
langchain-openai
msal
webbrowser